base for skills mostly done, missing Action

+++ include/ai/AI.h
@@ -6,13 +6,13 @@
 class AI {
   public:
     AI(){}
-    AI(long x, long y):origin_x(x),origin_y(y){}
+    AI(int x, int y):origin_x(x),origin_y(y){}
     virtual Action * getAction(Adventure * adventure, Character * c) = 0;
-    static int getFleeOrientation(Character * self, long x, long y);
-    static int getFollowOrientation(Character * self, long x, long y);
+    static int getFleeOrientation(Character * self, int x, int y);
+    static int getFollowOrientation(Character * self, int x, int y);
   protected:
-    long origin_x;
-    long origin_y;
+    int origin_x;
+    int origin_y;
 };
 
 #endif // _AI_H_
